+++ x/distribution/types/keys.go
@@ -17,9 +17,6 @@ const (
 
 	// RouterKey is the message route for distribution
 	RouterKey = ModuleName
-
-	// QuerierRoute is the querier route for distribution
-	QuerierRoute = ModuleName
 )
 
 // Keys for distribution store
