+++ x/mint/types/keys.go
@@ -13,9 +13,6 @@ const (
 	// StoreKey is the default store key for mint
 	StoreKey = ModuleName
 
-	// QuerierRoute is the querier route for the minting store.
-	QuerierRoute = StoreKey
-
 	// Query endpoints supported by the minting querier
 	QueryParameters       = "parameters"
 	QueryInflation        = "inflation"
