+++ x/params/types/keys.go
@@ -3,7 +3,4 @@ package types
 const (
 	// ModuleName defines the module name
 	ModuleName = "params"
-
-	// QuerierRoute defines the module's query routing key
-	QuerierRoute = ModuleName
 )
