+++ x/slashing/types/keys.go
@@ -17,9 +17,6 @@ const (
 
 	// RouterKey is the message route for slashing
 	RouterKey = ModuleName
-
-	// QuerierRoute is the querier route for slashing
-	QuerierRoute = ModuleName
 )
 
 // Keys for slashing store
