+++ x/staking/types/keys.go
@@ -21,9 +21,6 @@ const (
 	// StoreKey is the string store representation
 	StoreKey = ModuleName
 
-	// QuerierRoute is the querier route for the staking module
-	QuerierRoute = ModuleName
-
 	// RouterKey is the msg router key for the staking module
 	RouterKey = ModuleName
 )
