+++ x/upgrade/types/keys.go
@@ -11,9 +11,6 @@ const (
 
 	// StoreKey is the prefix under which we store this module's data
 	StoreKey = ModuleName
-
-	// QuerierKey is used to handle abci_query requests
-	QuerierKey = ModuleName
 )
 
 const (
